Rewrite and refactor fixtures -- esp. [DUMP.*].
Rewrite fixtures [DUMP.*] as [DUMP_*].  Then, Since [DUMP_dumpfiles]
relies on fixture [DUMP_dump], the former will fail if the latter is
not in its namespace. Rewrite the former so that it relies only on
[regtest].

Factor everything out of [testfixtures/__init__.py], leaving
it with only the role of importing things.